Pavo Pico Version 2 Settings and Configuration
These are the exact setting that are working in the most recent video. The  (Flip-Pavo-Pico-Version-2-Final-Settings.txt) file seems to do the trick for everything to work properly with your pavo pico version 2 (DJI 03) PNP version. 

Remember to bind your receiver and updated the firmware on the DJI FPV Remote Controller too through the DJI Assistant 2 for Consumer Drones.